Refactor: Split SavingsList into smaller components
Refactor SavingsList.tsx to improve maintainability by extracting functionalities into separate components and hooks. This includes creating components for the empty state, and a hook for Supabase listeners.
