+++ src/components/savings/hooks/useSavingsRealtimeUpdates.ts
@@ -1,0 +1,30 @@
+
+import { useEffect } from "react";
+import { supabase } from "@/integrations/supabase/client";
+import { useQueryClient } from "@tanstack/react-query";
+
+export const useSavingsRealtimeUpdates = () => {
+  const queryClient = useQueryClient();
+
+  useEffect(() => {
+    const channel = supabase
+      .channel('monthly-savings-changes')
+      .on(
+        'postgres_changes',
+        { 
+          event: '*',
+          schema: 'public',
+          table: 'monthly_savings'
+        },
+        (payload) => {
+          console.log('Modification des épargnes mensuelles:', payload);
+          queryClient.invalidateQueries({ queryKey: ['dashboard-data'] });
+        }
+      )
+      .subscribe();
+
+    return () => {
+      supabase.removeChannel(channel);
+    };
+  }, [queryClient]);
+};
